Fix design issue and always set unit_amount_rounded
The previous implementation relied on the override of _timesheet_postprocess
to compute unit_amount_rounded only if the line fulfils the condition that
were defined in _update_amount_rounded.
The issue is that this was done only for analytic line having a project_id
and the value stayed 0.0 for any other line.
However, as the qty_delivered on sale order line is computed using a read_group
we cannot ensure all the read lines will have a project_id, so if we replace
the value of unit_amount by unit_amount_rounded, we must make sure the value
is set even if the line is not linked to a project with a rounding policy.
Therefore unit_amount_rounded is now always updated, and is added to the
triggering fields of sale.order.line._compute_qty_delivered to make sure the
computation is correct.